# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Tekton Hub
  description: HTTP servie for maintaing Tetkon hub
  version: ""
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /resource/{resourceId}:
    get:
      tags:
      - resource
      summary: Info resource
      description: Get one Resource info
      operationId: resource#Info
      parameters:
      - name: resourceId
        in: path
        description: ID of resource to be shown
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ResourceInfoResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ResourceInfoInternalErrorResponseBody'
      schemes:
      - http
  /resources:
    get:
      tags:
      - resource
      summary: All resource
      description: Get all Resources
      operationId: resource#All
      responses:
        "200":
          description: OK response.
          schema:
            type: array
            items:
              $ref: '#/definitions/ResourceResponse'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ResourceAllInternalErrorResponseBody'
      schemes:
      - http
definitions:
  CatalogResponse:
    title: CatalogResponse
    type: object
    properties:
      id:
        type: integer
        description: ID is the unique id of the category
        example: 1
        format: int64
      type:
        type: string
        description: Type of support tier
        example: official
    example:
      id: 1
      type: official
    required:
    - id
    - type
  CatalogResponseBody:
    title: CatalogResponseBody
    type: object
    properties:
      id:
        type: integer
        description: ID is the unique id of the category
        example: 1
        format: int64
      type:
        type: string
        description: Type of support tier
        example: official
    example:
      id: 1
      type: official
    required:
    - id
    - type
  ResourceAllInternalErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: All_internal-error_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ResourceInfoInternalErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: Info_internal-error_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ResourceInfoResponseBody:
    title: 'Mediatype identifier: detail; view=default'
    type: object
    properties:
      catalog:
        $ref: '#/definitions/CatalogResponseBody'
      description:
        type: string
        description: Description of resource
        example: This Task is Golang task to build Go projects
      displayName:
        type: string
        description: Display name of the resource
        example: golang
      id:
        type: integer
        description: ID is the unique id of the resource
        example: 1
        format: int64
      last_updated_at:
        type: string
        description: Date when resource was last updated
        example: "2020-03-26T14:09:08.19599+05:30"
      latest_version:
        type: string
        description: Latest version o resource
        example: "0.1"
      name:
        type: string
        description: Name of the resource
        example: golang
      rating:
        type: integer
        description: Rating of resource
        example: 5
        format: int64
      type:
        type: string
        description: Type of resource
        example: task
      versions:
        type: array
        items:
          $ref: '#/definitions/VersionsResponseBody'
        description: Version of resource
        example:
        - version: 0,1
          versionId: 1
        - version: 0,1
          versionId: 1
    description: InfoResponseBody result type (default view)
    example:
      catalog:
        id: 1
        type: official
      description: This Task is Golang task to build Go projects
      displayName: golang
      id: 1
      last_updated_at: "2020-03-26T14:09:08.19599+05:30"
      latest_version: "0.1"
      name: golang
      rating: 5
      type: task
      versions:
      - version: 0,1
        versionId: 1
      - version: 0,1
        versionId: 1
      - version: 0,1
        versionId: 1
      - version: 0,1
        versionId: 1
    required:
    - id
    - name
    - displayName
    - catalog
    - type
    - description
    - latest_version
    - rating
    - last_updated_at
    - versions
  ResourceResponse:
    title: ResourceResponse
    type: object
    properties:
      catalog:
        $ref: '#/definitions/CatalogResponse'
      description:
        type: string
        description: Description of resource
        example: This Task is Golang task to build Go projects
      displayName:
        type: string
        description: Display name of the resource
        example: golang
      id:
        type: integer
        description: ID is the unique id of the resource
        example: 1
        format: int64
      last_updated_at:
        type: string
        description: Date when resource was last updated
        example: "2020-03-26T14:09:08.19599+05:30"
      latest_version:
        type: string
        description: Latest version o resource
        example: "0.1"
      name:
        type: string
        description: Name of the resource
        example: golang
      rating:
        type: integer
        description: Rating of resource
        example: 5
        format: int64
      tags:
        type: array
        items:
          $ref: '#/definitions/TagResponse'
        description: Tags related to resources
        example:
        - id: 1
          name: notification
        - id: 1
          name: notification
        - id: 1
          name: notification
      type:
        type: string
        description: Type of resource
        example: task
    example:
      catalog:
        id: 1
        type: official
      description: This Task is Golang task to build Go projects
      displayName: golang
      id: 1
      last_updated_at: "2020-03-26T14:09:08.19599+05:30"
      latest_version: "0.1"
      name: golang
      rating: 5
      tags:
      - id: 1
        name: notification
      - id: 1
        name: notification
      - id: 1
        name: notification
      type: task
    required:
    - id
    - name
    - displayName
    - catalog
    - type
    - description
    - latest_version
    - rating
    - last_updated_at
    - tags
  TagResponse:
    title: TagResponse
    type: object
    properties:
      id:
        type: integer
        description: ID is the unique id of the tag
        example: 1
        format: int64
      name:
        type: string
        description: Name of the tag
        example: notification
    example:
      id: 1
      name: notification
    required:
    - id
    - name
  VersionsResponseBody:
    title: VersionsResponseBody
    type: object
    properties:
      version:
        type: string
        description: Version of the resource to be fetched
        example: 0,1
      versionId:
        type: integer
        description: Version ID of the resource to be fetched
        example: 1
        format: int64
    example:
      version: 0,1
      versionId: 1
    required:
    - versionId
    - version